+++ src/components/CurrencyInputPanel/index.js
@@ -1,5 +1,4 @@
 import React, { useState, useRef, useMemo } from 'react'
-import { Link } from 'react-router-dom'
 import { useTranslation } from 'react-i18next'
 import { ethers } from 'ethers'
 import { BigNumber } from '@uniswap/sdk'
